:sparkles: Add username to localstorage in AddUser.tsx
